+++ server/entity/Serializable.py
